영호가 철진이에게 주는 작은 선물 (댓글 삭제 확인기능) -영호

+++ frontend/src/containers/StoryDetail.js
@@ -506,9 +506,12 @@ const StoryDetail = ({ match }) => {
   };
 
   const commentDeleteHandler = (id) => {
-    dispatch(commentDelete({ comment_id: id, story_id: data.id })).then(
-      setComments("")
-    );
+    const ok = window.confirm("삭제하시겠습니까?")
+    if(ok){
+      dispatch(commentDelete({ comment_id: id, story_id: data.id })).then(
+          setComments("")
+      );
+    }
   };
 
   const commentClear = () => {
